Significant change to joinWith()
Instead of joinWith('LEFT', 'relatedFieldName'), it is now joinWith('LEFT relatedFieldName'). If you specify only a relatedFieldName, the join type will be JOIN, just as with select->join(). You can specify an alias using AS, a la joinWith('LEFT relatedFieldName AS alias_name'). You can also pass a callable of the form function ($sub) as an optional second param to do a sub-joinWith on the related, allowing you to chain foreign relationships.